Preliminary task: Moved DTO SememeSlim from oapenwb-javalin to oapenwb-persistency, and removed link to LemmaTemplateProcessor from entity LemmaTemplate.
